started refactoring search
